Some QA adjustments
