+++ jwt-server/spring/jwt-spring-auth0-webflux/src/main/java/org/entur/jwt/spring/reactive/auth0/Auth0SecurityAutoConfigurationWebFlux.java
@@ -2,7 +2,6 @@ package org.entur.jwt.spring.reactive.auth0;
 
 import org.entur.jwt.spring.Auth0SecurityAutoConfiguration;
 import org.entur.jwt.spring.JwtAutoConfigurationWebFlux;
-import org.entur.jwt.spring.filter.JwtAuthenticationManager;
 import org.entur.jwt.spring.filter.resolver.JwtArgumentResolver;
 import org.entur.jwt.spring.filter.resolver.JwtPayload;
 import org.entur.jwt.spring.properties.Auth0Properties;
@@ -12,7 +11,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.ReactiveAuthenticationManager;
 
 @Configuration
 @EnableConfigurationProperties(Auth0Properties.class)
@@ -24,13 +22,6 @@ public class Auth0SecurityAutoConfigurationWebFlux extends Auth0SecurityAutoConf
     @ConditionalOnProperty(name = "entur.jwt.enabled", havingValue = "true")
     public JwtArgumentResolver jwtArgumentResolver() {
         return new JwtArgumentResolver((value, type) -> new JwtPayload(value), JwtPayload.class);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(ReactiveAuthenticationManager.class)
-    @ConditionalOnProperty(name = "entur.jwt.enabled", havingValue = "true")
-    public ReactiveAuthenticationManager reactiveAuthenticationManager() {
-        return new JwtAuthenticationManager();
     }
 
 }
