+++ jwt-server/spring/jwt-spring-web/src/main/java/org/entur/jwt/spring/JwtAutoConfigurationWeb.java
@@ -1,9 +1,14 @@
 package org.entur.jwt.spring;
 
-import org.entur.jwt.spring.filter.*;
+import org.entur.jwt.spring.filter.JwtAuthenticationExceptionAdvice;
+import org.entur.jwt.spring.filter.JwtAuthenticationFilter;
+import org.entur.jwt.spring.filter.JwtAuthorityMapper;
+import org.entur.jwt.spring.filter.JwtDetailsMapper;
+import org.entur.jwt.spring.filter.JwtPrincipalMapper;
 import org.entur.jwt.spring.filter.log.JwtMappedDiagnosticContextMapper;
 import org.entur.jwt.spring.filter.resolver.JwtArgumentResolver;
 import org.entur.jwt.spring.properties.AuthorizationProperties;
+import org.entur.jwt.spring.properties.CorsProperties;
 import org.entur.jwt.spring.properties.PermitAll;
 import org.entur.jwt.spring.properties.SecurityProperties;
 import org.entur.jwt.verifier.JwtClaimExtractor;
@@ -16,10 +21,15 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.HandlerExceptionResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 @ConditionalOnProperty(name = {"entur.jwt.enabled"}, havingValue = "true")
@@ -65,6 +75,74 @@ public class JwtAutoConfigurationWeb extends JwtAutoConfiguration {
     @ConditionalOnMissingBean(JwtAuthenticationExceptionAdvice.class) // allow for customization
     public JwtAuthenticationExceptionAdvice advice() {
         return new JwtAuthenticationExceptionAdvice();
+    }
+
+    @Bean("corsConfigurationSource")
+    @ConditionalOnProperty(name = {"entur.cors.enabled"}, havingValue = "true")
+    public CorsConfigurationSource corsConfigurationSource(SecurityProperties oidcAuthProperties) {
+        CorsProperties cors = oidcAuthProperties.getCors();
+        if (cors.getMode().equals("api")) {
+            return getCorsConfiguration(cors);
+        } else {
+            if (!cors.getOrigins().isEmpty()) {
+                throw new IllegalStateException("Expected empty hosts configuration for CORS mode '" + cors.getMode() + "'");
+            }
+            log.info("Disable CORS requests for webapp mode");
+
+            return getEmptyCorsConfiguration();
+        }
+    }
+
+    @Bean("corsConfigurationSource")
+    @ConditionalOnProperty(name = {"entur.security.cors.mode"}, havingValue = "webapp")
+    public CorsConfigurationSource corsConfigurationSourceForWebapp(SecurityProperties properties) {
+        log.info("Disable CORS requests for webapp mode");
+        return getEmptyCorsConfiguration();
+    }
+
+    public static CorsConfigurationSource getEmptyCorsConfiguration() {
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowedOrigins(Collections.emptyList());
+        config.setAllowedHeaders(Collections.emptyList());
+        config.setAllowedMethods(Collections.emptyList());
+
+        source.registerCorsConfiguration("/**", config);
+
+        return source;
+    }
+
+    public static CorsConfigurationSource getCorsConfiguration(CorsProperties properties) {
+        List<String> defaultAllowedMethods = Arrays.asList("GET", "HEAD", "POST", "PUT", "DELETE", "PATCH", "OPTIONS");
+        List<String> defaultAllowedHeaders = Collections.singletonList("*");
+
+        List<String> origins = properties.getOrigins();
+        log.info("Enable CORS request with origins {}, methods {} and headers {} for API mode",
+                properties.getOrigins(),
+                properties.hasMethods() ? properties.getMethods() : "default (" + defaultAllowedMethods + ")",
+                properties.hasHeaders() ? properties.getHeaders() : "default (" + defaultAllowedHeaders + ")"
+        );
+
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowedOrigins(origins);
+        if (properties.hasHeaders()) {
+            config.setAllowedHeaders(properties.getHeaders());
+        } else {
+            config.setAllowedHeaders(defaultAllowedHeaders);
+        }
+        if (properties.hasMethods()) {
+            config.setAllowedMethods(properties.getMethods());
+        } else {
+            config.setAllowedMethods(defaultAllowedMethods); // XXX
+        }
+        config.setMaxAge(86400L);
+        config.setAllowCredentials(true);
+        source.registerCorsConfiguration("/**", config);
+
+        return source;
     }
 
 }
