+++ jwt-server/spring/jwt-spring-webflux/src/main/java/org/entur/jwt/spring/CustomAuthenticationFailureHandler.java
@@ -1,0 +1,19 @@
+package org.entur.jwt.spring;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.server.reactive.ServerHttpResponse;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.server.WebFilterExchange;
+import org.springframework.security.web.server.authentication.ServerAuthenticationFailureHandler;
+import reactor.core.publisher.Mono;
+
+public class CustomAuthenticationFailureHandler implements ServerAuthenticationFailureHandler {
+
+    @Override
+    public Mono<Void> onAuthenticationFailure(WebFilterExchange exchange, AuthenticationException exception) {
+        return Mono.fromRunnable(() -> {
+            ServerHttpResponse response = exchange.getExchange().getResponse();
+            response.setStatusCode(HttpStatus.FORBIDDEN);
+        });
+    }
+}
